aggregated optimized control tests into one file
